requirements.txt
fastapi
pydantic
uvicorn
watchfiles
oauth2-client
python-dotenv
slack_sdk
slack_bolt